refactor(hickey): isolate board placeholder leaf
Move the temporary board UI out of AppRoutes so the route module focuses on path selection.
